requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
python-dotenv==1.0.1
httpx==0.26.0
outscraper==5.9.1
requests==2.31.0
apify-client
openai==1.35.0
serpapi==0.1.5 